Changed raw primitive data to match methods parameters on SDK

+++ app/src/main/java/spotter/netty/org/nettyspotter/rawmodel/Geometry.kt
@@ -9,5 +9,5 @@ data class Geometry(
         var type: String,
         @SerializedName("coordinates")
         @Expose
-        var coordinates: List<Float>
+        var coordinates: List<Double>
 )
